requirements.txt
typer
pydantic
PyYAML
pytest
numpy
